+++ apps/toolkit-app/next.config.js
@@ -1,10 +1,15 @@
+// @ts-check
+
 const withBundleAnalyzer = require("@next/bundle-analyzer")({
   enabled: process.env.ANALYZE === "true",
 })
 const { withBlitz } = require("@blitzjs/next")
 
-module.exports = withBlitz(
-  withBundleAnalyzer({
-    reactStrictMode: true,
-  })
-)
+/**
+ * @type {import('@blitzjs/next').BlitzConfig}
+ **/
+const config = {
+  reactStrictMode: true,
+}
+
+module.exports = withBlitz(withBundleAnalyzer(config))